// src/types.ts
// Enum for spinner game options
export enum GameOption {
    SurpriseEasy = 'Surprise Easy Game',
    MathFun = 'Math Fun',
    TypingSpeed = 'Typing Speed',
    Matching = 'Matching',
}

// Enum for merge conflict resolution options
export enum MergeOption {
    Current = 'Current Changes',
    Incoming = 'Incoming Changes',
    Combination = 'Combination',
}